feat(#1737): Move OAuth2 LoginModule to dedicated activemq-oauth2 module
Extract OAuth2 JWT authentication into its own activemq-oauth2 module
(similar to activemq-shiro) with package org.apache.activemq.oauth2.
Remove nimbus-jose-jwt dependency from activemq-jaas and assembly,
add activemq-oauth2 to parent modules/dependencyManagement and assembly.
Include commented-out OAuth2 configuration example in the distribution
login.config.
