+++ activemq-jaas/pom.xml
@@ -106,11 +106,6 @@
       <artifactId>jasypt</artifactId>
       <optional>true</optional>
     </dependency>
-    <dependency>
-      <groupId>com.nimbusds</groupId>
-      <artifactId>nimbus-jose-jwt</artifactId>
-      <optional>true</optional>
-    </dependency>
   </dependencies>
 
   <profiles>
