+++ activemq-jaas/src/test/resources/login.config
@@ -179,16 +179,6 @@ GuestLoginWithDefaults {
         debug=true;
 };
 
-OAuth2Login {
-    org.apache.activemq.jaas.OAuth2LoginModule required
-        debug=true
-        oauth2.jwksUrl="https://idp.example.com/.well-known/jwks.json"
-        oauth2.issuer="https://idp.example.com"
-        oauth2.audience="activemq"
-        oauth2.usernameClaim="sub"
-        oauth2.groupsClaim="groups";
-};
-
 OpenLdapConfiguration {
   org.apache.activemq.jaas.LDAPLoginModule required
         debug=true
