+++ activemq-oauth2/src/main/java/org/apache/activemq/oauth2/OAuth2TokenCallback.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.activemq.jaas;
+package org.apache.activemq.oauth2;
 
 import javax.security.auth.callback.Callback;
