+++ assembly/src/release/conf/login.config
@@ -19,3 +19,30 @@ activemq {
         org.apache.activemq.jaas.properties.user="users.properties"
         org.apache.activemq.jaas.properties.group="groups.properties";
 };
+
+/**
+ * OAuth2 JWT authentication example.
+ * Uncomment and configure the following to authenticate clients using
+ * OAuth2/OIDC JWT access tokens. The token is passed as the password
+ * in the connection credentials.
+ *
+ * Required options:
+ *   oauth2.jwksUrl  - URL to the JWKS endpoint for token signature verification
+ *   oauth2.issuer   - Expected token issuer (iss claim)
+ *
+ * Optional options:
+ *   oauth2.audience      - Expected token audience (aud claim)
+ *   oauth2.usernameClaim - JWT claim to extract the username from (default: "sub")
+ *   oauth2.groupsClaim   - JWT claim containing group memberships (default: "groups")
+ *
+ * To use this configuration, change the broker's jaasAuthenticationPlugin
+ * to reference "activemq-oauth2" instead of "activemq".
+ */
+//activemq-oauth2 {
+//    org.apache.activemq.oauth2.OAuth2LoginModule required
+//        oauth2.jwksUrl="https://idp.example.com/.well-known/jwks.json"
+//        oauth2.issuer="https://idp.example.com"
+//        oauth2.audience="activemq"
+//        oauth2.usernameClaim="preferred_username"
+//        oauth2.groupsClaim="roles";
+//};
